Update to use ember-data 3.0
Ember Data 3.0 removed the ContainerInstanceCache class which we were
(prior to this patch) monkey-patching to override the default serailizer
fallback lookup.

This PR goes back to a similar approach to what
ember-data-model-fragments was using prior to ember-data 2.11 (when
ContainerInstanceCache was added)

For a bit of context here is the patch that introduced the monkey-patch
for Ember Data 2.11:

https://github.com/lytics/ember-data-model-fragments/commit/a856c24d82f0f655fc21474e83b2f9bf9a605f5a

And here is the patch that remove ContainerInstanceCache class from
Ember Data making the aforementioned monkey patch throw in 3.0:

https://github.com/emberjs/data/pull/5300

This PR ensures that the correct serializer is looked up as per the
tests.