Inject store into transforms
Ideally we'd be more targeted with the injection of store, but I
couldn't figure out how to do that and still allow it to work.

We rely on store being present in our transforms, but Ember Data nor
Ember itself inject the store into transforms.  So, I'm not entirely
certain how this ever worked.

Here is an issue that looks remarkably similar to the problem I ran
into.  Though it doesn't seem resolved:

https://github.com/lytics/ember-data-model-fragments/issues/263

Also, remove transform:fragment/fragment-array/array registrations which
are not necessary becuase they are in app/