Prevent infinite recursion on save
Prevent infinite recursion when saving a data containing a fragment which contains a fragment which contains a fragmentArray which has changed due to the way notifyPropertyChange works. setupFragment was calling setFragmentData which cause setupFragment to be called again, and never give it a change to assign fragment in internal model data.