[breaking][bugfix] respect `null` on server updates
Prior to this change, if a `fragment` was initialized with some value,
and then server responds with a `null` it did not have any impact
on a fragment state, and it kept having a non-empty value assigned.
Now it's fixed, so the fragment attribute can be updated to `null`.

The same behavior is applied to the `fragmentArray`.
Technically, this is a *breaking change*, since before, if non empty
`fragmentArray` received a `null` from the server,
instead of nullifying the attribute, the `fragmentArray` instance just
became empty(with no items inside).
It was misaligned with how the `fragmentArray` handles `null` on create,
cause during creation it is nullable. So for consistency,
we let it be nullable on updates as well.